requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
alembic
bcrypt==4.0.1
passlib[bcrypt]
python-jose
python-dotenv~=1.0.1
python-multipart
pydantic~=2.10.6
pytz
pytest-cov
pytest
coverage
httpx
locust